updated: enemy hitboxes, enemy rendering, enemy attack; fixed some bugs
enemies now all have the pivot on bottom of the sprite to assure proper rendering and to assure knockback always happens in the intuitive direction.

enemy healthbars are now shown layer2, which means they are above everything else and are always visible.

enemies now inflict the damage on the frame on which it makes the most sense instead of at the begging of the attack animation. this also has the effect that killing an enemy before it inflicts damage is possible even after the attack animation started.

fixed enemies attacking multiple times instead of only once.
fixed player sometimes not showing deathanimation after hp <= 0
fixed bug with trees beeing hidden behind walls